taijios-lite/requirements.txt
python-docx
openai>=1.0.0
python-dotenv
aiohttp>=3.9.0
